Boiler plate for commands
This adds a bunch of code for setting up the commands we will need to
implement to get in line with the ruby client. I wanted to use the flags
package but it would have added a bunch of ugly code to do the same
thing as urfave/cli does.