Better testing and formatter checking
This mostly just adds in a better way to test all of these functions by
wrapping them inside the context of a cli command being run. Everything
now is as we expect and we can just write out commands as we would on
the cli for testing. This also adds a small script to check what format
we should be outputting in.