Add basic CQL support
This adds some basic CQL generation support. It additionally adds the
option to switch between joinin with and or or which looks like it was
meant to be a feature of the ruby client but never made its way in. It's
honestly not all the useful but might come in useful.