Add basic attr setting support
This adds some basic ability to pipe to the modify/set command and have
it set attributes. The code here is looking a little more dense than I
would like but the collins set command has a log of logic packed into it
that isn't easy to make look nice in golang.