Add hostname arg support for query
This adds support for passing in the hostname as the first non plag
param after collins query. This is a little messy though since it looks
like the ruby client handles this differently and maybe just builds it's
own CQL without the use of the Attribute field we are using. It would
likely be best to refactor to not use the Attribute field at all and
lower the complexity a bit.