Add GPU info to collins query
This gives the ability to get the information about GPUs off a given
asset. Currently it just pulls the info from the first asset which for
most usecases should be fine. If needed we can add an ability to specify
which GPU you want to output for in the future.