Support piping to query cmd
This lets you take a tags from a list of files and pipe it to collins
query to get more info about them. Only the tags flag or stdin is
currently supported. With some more work it can accept both but the CQL
is going to need to be refactored a little bit more than i'm willing to
do at this second.