Fix piping for huge files
Collins by default has a sane 4096 bytes allowed to be passed to it via
a URL. Since these files in practice can be very long sometimes you can
easily cause collins to blow up. When it overflows 4096 bytes it doesn't
return anything at all to the client. This limits the size of requests
we send to it by a lot but also adds the ability to change the number of
keys we include in the CQL. We will let the operator choose what size to
use for now.