Add better negative matching
This switches the way that we do negative matching so it's consistent.
It's always a ~ before the value. in attributes it's key:~value and for
other flags it's always ~something. This is largely useful only for
attributes and filtering on lists passed into collins query although i'm
sure it has limited use for other things as well.