Set attributes on provision
Move the ability to set attributes into provision so you don't need to
pipe to an additional `collins set` command. In the case where you are
provisioning lots of hosts this might not work well anyway since you have
to wait a long time for all hosts to finish before the attribute would be
set.